+++ src/lib.rs
@@ -42,6 +42,7 @@ pub use reachability::{FunctionImplementation, Reachability, ReachableCall, Reac
 mod scope;
 mod services;
 mod stack_trace;
+mod string_interning;
 mod string_value;
 pub use string_value::StringValue;
 pub use services::{MissingService, RuntimeService, ServiceUse};